+++ packages/authentication/modules/user/index.ts
@@ -1,5 +1,7 @@
 export { default as getUser } from './getUser'
 export { default as useSimpleTokenUser } from './useSimpleTokenUser'
+export { default as useUpdateUser } from './useUpdateUser'
+export { default as useJWTUser } from './useJWTUser'
 export { default as useUser } from './useUser'
 
 export { default as withUser } from './withUser'