+++ packages/authentication/modules/user/useUser/index.ts
@@ -8,7 +8,12 @@ import { IUseUser } from './types'
 
 /**
  * @deprecated
- * Prefer using the `getUser` function instead.
+ * On the server side:
+ * - Prefer using the `getUser` function.
+ *
+ * On the client side:
+ * - If you are using `JWT`, prefer using the `useJWTUser` hook.
+ * - If you are using `Simple Token`, prefer using the `useSimpleTokenUser` hook.
  */
 const useUser = <TUser extends Partial<IUser> & Partial<IJWTContent>>({
   cookieName = ACCESS_COOKIE_NAME,