+++ packages/authentication/services/user.ts
@@ -1,10 +1,17 @@
-import { axios } from '@baseapp-frontend/utils'
+import { axios, axiosForFiles } from '@baseapp-frontend/utils'
 
-import { IUser } from '../types/user'
+import { IUser, UserUpdateParams } from '../types/user'
 
 export default class UserApi {
   static getUser<TUser extends Partial<IUser>>(): Promise<TUser> {
     return axios.get(`/users/me`)
+  }
+
+  static updateUser<TUser extends Partial<IUser>>({
+    userId,
+    data,
+  }: UserUpdateParams<TUser>): Promise<TUser> {
+    return axiosForFiles.patch(`/users/${userId}`, data)
   }
 }