+++ packages/utils/functions/form/index.ts
@@ -1,4 +1,6 @@
 export { default as withController } from './withController'
 export type * from './withController/types'
 
+export * from './filterDirtyValues'
+export * from './getInitialValues'
 export * from './setFormApiErrors'